third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-basic.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Test if requestMIDIAccess is available.");

shouldBeDefined("navigator.requestMIDIAccess");

var promise = navigator.requestMIDIAccess();
shouldBeDefined("promise");

shouldBeDefined("promise.then");

promise.then(function(){}, function(){});

window.successfullyParsed = true;

</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
